+++ Changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.0] - ?????
 ### Added
 - `Base.get` and `Base.getindex` are now both implemented for `Const` (like they are implemented already for `Identity`), to simplify working with `Const`.
+- Vector conversions (`convert`) are now more generic and support AbstractArray in general 
 
 ### Fixed
 - `Option(3)` works now